Add deserialize-4
The idea here is to provide functions in TypeScript and a tree
of decoder functions which describe the expected input.

The morphir-elm tool then only has to generate the tree of
decoders, it doesn't generate any actual functions.

I got muddled between two different approaches here, there are
two options for specifying the schema.

  1. A tree of decoder functions, using `bind()` to set parameters
     on arguments that define the structure.

  2. A tree of Schema.Type {} instances, which could be used to
     generate the above.

I should switch to approach 1 and if it sucks, try approach 2.